Major Code Organization
I reorganized a lot of the code to a more readable format. This will also be pushed to the main branch as this is the format I want to keep as we move forward.
