--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FSM/State.java
@@ -1,15 +1,0 @@
-package org.firstinspires.ftc.teamcode.FSM;
-
-public abstract class State {
-    public State(){
-
-    }
-
-    public abstract void state_action();
-    public void do_action(){
-        state_action();
-        return;
-    }
-
-
-}
