+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TEST/IMU_TESTER.java
@@ -2,17 +2,9 @@ package org.firstinspires.ftc.teamcode.TEST;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.HardwareMap.HMap;
-import org.firstinspires.ftc.teamcode.HardwareMap.IMUPlus;
-import org.firstinspires.ftc.teamcode.HardwareMap.MotorPP;
-import org.firstinspires.ftc.teamcode.HardwareMap.MotorPlus;
-
-import static org.firstinspires.ftc.teamcode.MISC.CONSTANTS.aPid_;
-import static org.firstinspires.ftc.teamcode.MISC.CONSTANTS.dPid_;
-import static org.firstinspires.ftc.teamcode.MISC.CONSTANTS.imuPid_;
-import static org.firstinspires.ftc.teamcode.MISC.CONSTANTS.vPid_;
+import org.firstinspires.ftc.teamcode.middleend.HardwareMappings.HMap;
+import org.firstinspires.ftc.teamcode.backend.hardware_extensions.MotorPlus;
 
 @Autonomous(name = "IMU PID Tester")
 public class IMU_TESTER extends LinearOpMode {
