--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOps/TeleV1.java
@@ -1,14 +1,0 @@
-package org.firstinspires.ftc.teamcode.TeleOps;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-class TeleV1 extends LinearOpMode {
-
-    @Override
-    public void runOpMode() throws InterruptedException {
-        waitForStart();
-        while(opModeIsActive() && !isStopRequested()){
-
-        }
-    }
-}
